# Improve batch send performances